Updates the css/styles.css page with the new Linter and added a new design in the joaobm-contact.html
If applied, this commit will have updated all the misspelings, abscense of ambivalent syntax  with a more readable code and consistent code style through the Linter 'stylelint'.

Was as well changed the contact page design giving a new dotted box around the email, telephone and message. As well the color of the same elements have been changed.